[step3] AutoTicketOffice, ManualTicketOffice 클래스 추가
기존 TicketOffice 클래스를 상속하는 AutoTicketOffice, ManualTicketOffice 클래스 추가.
각 클래스가 issueTickets()를 오버라이딩하는 형태로 구현.
