+++ src/main/java/domain/LottoTicket.java
@@ -41,5 +41,4 @@ public class LottoTicket {
     }
 
 
-
 }
